Implement Addrinfo#initialize based on MRI
This setup is a _very_ rough setup based on the behaviour of MRI. I'm
still not sure everything is correct, but at least both Rubinius and MRI
now pass all specs.

This commit adds a dependency on rubysl-resolv. Resolv exposes
Resolv::IPv6::Regex which is used for validating IPv6 addresses should
the address family of an Addrinfo be set to AF_INET6. MRI seems to
delegate this to its getaddrinfo() implementation, but we might as well
just use a regex for this (saving a system call).