Disable AncillaryData#data for ip(v6)_pktinfo
On MRI Socket::AncillaryData#data returns a String representation of the
underlying in_pktinfo/in6_pktinfo C structure. This is problematic
because of the following:

1. The structure _might_ differ on platforms, thus leading to different
   (possibly incompatible) output.
2. While this data can be mimiced for IPv4 it's a total pain to do for
   IPv6, meaning one _has_ to also use these structures under the hood.
3. On Linux the in6_pktinfo structure is only available when the macro
   "__USE_GNU" is defined, typically done by defining "_GNU_SOURCE".
   Defining _GNU_SOURCE in turn breaks a lot of parts of Rubinius for
   reasons unknown (e.g. random types not being visible and segfaults
   as a result).

This basically left us with the following options:

1. Mimic MRI internals in pure Ruby, possibly using a lot of hacks to
   get this to work for IPv6 (getting this to work for IPv4 was already
   gross).
2. Using the same C structures, possibly having to spend weeks to get
   Rubinius to compile with _GNU_SOURCE defined.
3. Not supporting AncillaryData#data when an object was created using
   either AncillaryData.ip_pktinfo or AncillaryData.ipv6_pktinfo.

Having discussed this with @brixen & friends we decided to go with
option 3. After searching for the "ip_pktinfo" and "ipv6_pktinfo"
methods on GitHub I can't find a single instance where these methods are
used in combination with the "data" method. Using the raw "data" method
is also rather pointless to begin with as the "ip_pktinfo" and
"ipv6_pktinfo" instance methods return the objects you'd want to work
with 99.99% of the time.

Dropping support for "data" means we can simply set an instance variable
containing the data we need, and duplicating this upon calling the
pktinfo instance methods (since MRI returns new Addrinfo objects upon
every call).