Wrapped up support for BasicSocket#recvmsg
The current implementation does _not_ support ancillary data. The extra
arguments used for setting the maximum length and the "scm_rights"
option are ignored, the return value will never include ancillary data.

Ancillary data is hard to support as reading/generating the data
requires the usage of C macros that might work in different ways on
different platforms. The only way to use these macros would be to define
custom C functions somewhere and bind those to Ruby using our FFI API.
This conflicts with the idea of not wanting to write extra C code as
well as further complicating the Rubinius build system and this Gem
itself.

Because ancillary data has a fairly limited use case and I was unable to
find any popular projects using it I've decided to _not_ support it for
the time being. Socket::AncillaryData will continue to exist for the
time being as it's useful to keep around (in case we ever do want to
support ancillary data) and because other methods might need it.