Removed ListenAndAccept
This module has been broken up in to two new modules:

* RubySL::Socket::Listen
* RubySL::Socket::Accept

Both contain class methods that can be used to call listen(), accept(),
and other related functions. This new setup also relies on explicitly
specifying what classes to use for new sockets, instead of relying on
the parent class of a source socket.