Revamp Socket.socketpair
This ensures both the address family and socket type are converted
properly (instead of raising an error which then raises another error
due to a typo). This also removes the 4th "klass" argument of
Socket.socketpair as said argument is not available in CRuby, nor it is
needed either.

Finally some FFI code has been moved to the RubySL::Socket::Foreign
module to make it easier to re-use.