Fixed free'ing of ifaddrs structs
The C structures contain various pointers that can't be passed to
free(), most likely because they're pointers to static memory. Rubinius
in turn tries to free these pointers the moment they go out of scope as
"autorelease" is enabled by default (including for regular FFI::Pointer
instances). This in turn would lead to segmentation faults whenever the
finalizer thread would make an attempt to free() invalid memory.

To work around this the autorelease option is disabled for all pointers
stored in an Ifaddrs instance, including the pointer of the structure
itself.