add short sleep in loop so this works on OSX
The spec using this helper (tcp_server_loop_spec) was hanging
indefinitely on OSX. It didn't matter what timeout value was used
for the helper. By adding a short sleep, the spec no longer hangs.
I assume a bug in rubinius but I do not know how to test for it
or solve it.