fix two small bugs
1. On some platforms like FreeBSD, the code to binary pack a string
while building a SockaddrUn struct did not work properly. So we now
explicitly set the fields :sun_family and :sun_path to the proper
values. This ensures they are laid out in memory properly for each
platform.

2. As a side effect of 1, an empty SockaddrUn instance would return
nil for the :sun_path field. We expect a string so we call :to_s on
any data returned to ensure it is a string.